toolkit/components/extensions/test/mochitest/test_ext_webrequest_auth.html

<!DOCTYPE HTML>

<html>
<head>
<meta charset="utf-8">
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SpawnTask.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/ExtensionTestUtils.js"></script>
  <script type="text/javascript" src="head_webrequest.js"></script>
  <script type="text/javascript" src="head.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
<script>
"use strict";

let baseUrl = "http://mochi.test:8888/tests/toolkit/components/passwordmgr/test/authenticate.sjs";
function testXHR(url) {
  return new Promise((resolve, reject) => {
    let xhr = new XMLHttpRequest();
    xhr.open("GET", url);
    xhr.onload = resolve;
    xhr.onabort = reject;
    xhr.onerror = reject;
    xhr.send();
  });
}

function getAuthHandler(result, blocking = true) {
  function background(result) {
    browser.webRequest.onAuthRequired.addListener((details) => {
      browser.test.succeed(`authHandler.onAuthRequired called with ${details.requestId} ${details.url} result ${JSON.stringify(result)}`);
      browser.test.sendMessage("onAuthRequired");
      return result;
    }, {urls: ["*://mochi.test/*"]}, ["blocking"]);
    browser.webRequest.onCompleted.addListener((details) => {
      browser.test.succeed(`authHandler.onCompleted called with ${details.requestId} ${details.url}`);
      browser.test.sendMessage("onCompleted");
    }, {urls: ["*://mochi.test/*"]});
    browser.webRequest.onErrorOccurred.addListener((details) => {
      browser.test.succeed(`authHandler.onErrorOccurred called with ${details.requestId} ${details.url}`);
      browser.test.sendMessage("onErrorOccurred");
    }, {urls: ["*://mochi.test/*"]});
  }

  let permissions = [
    "webRequest",
    "*://mochi.test/*",
  ];
  if (blocking) {
    permissions.push("webRequestBlocking");
  }
  return ExtensionTestUtils.loadExtension({
    manifest: {
      permissions,
    },
    background: `(${background})(${JSON.stringify(result)})`,
  });
}

add_task(function* test_webRequest_auth() {
  // Make use of head_webrequest to ensure event sequence.
  let events = {
    "onBeforeRequest":     [{urls: ["*://mochi.test/*"]}, ["blocking"]],
    "onBeforeSendHeaders": [{urls: ["*://mochi.test/*"]}, ["blocking", "requestHeaders"]],
    "onSendHeaders":       [{urls: ["*://mochi.test/*"]}, ["requestHeaders"]],
    "onBeforeRedirect":    [{urls: ["*://mochi.test/*"]}],
    "onHeadersReceived":   [{urls: ["*://mochi.test/*"]}, ["blocking", "responseHeaders"]],
    "onAuthRequired":      [{urls: ["*://mochi.test/*"]}, ["blocking", "responseHeaders"]],
    "onResponseStarted":   [{urls: ["*://mochi.test/*"]}],
    "onCompleted":         [{urls: ["*://mochi.test/*"]}, ["responseHeaders"]],
    "onErrorOccurred":     [{urls: ["*://mochi.test/*"]}],
  };

  let extension = makeExtension(events);
  yield extension.startup();
  let authInfo = {
    username: "testuser",
    password: "testpass",
  };
  let expect = {
    "authenticate.sjs": {
      type: "xmlhttprequest",
      // we expect these additional events after onAuthRequired
      optional_events: ["onBeforeRequest", "onHeadersReceived"],
      authInfo,
    },
  };
  // expecting origin == undefined
  extension.sendMessage("set-expected", {expect, origin: location.href});
  yield extension.awaitMessage("continue");

  yield testXHR(`${baseUrl}?realm=webRequest_auth&user=${authInfo.username}&pass=${authInfo.password}`);

  yield extension.awaitMessage("done");
  yield extension.unload();
});

// This test is the same as above, however we shouldn't receive onAuthRequired
// since those credentials are now cached (thus optional_events is not set).
add_task(function* test_webRequest_cached_credentials() {
  // Make use of head_webrequest to ensure event sequence.
  let events = {
    "onBeforeRequest":     [{urls: ["*://mochi.test/*"]}, ["blocking"]],
    "onBeforeSendHeaders": [{urls: ["*://mochi.test/*"]}, ["blocking", "requestHeaders"]],
    "onSendHeaders":       [{urls: ["*://mochi.test/*"]}, ["requestHeaders"]],
    "onBeforeRedirect":    [{urls: ["*://mochi.test/*"]}],
    "onHeadersReceived":   [{urls: ["*://mochi.test/*"]}, ["blocking", "responseHeaders"]],
    "onAuthRequired":      [{urls: ["*://mochi.test/*"]}, ["blocking", "responseHeaders"]],
    "onResponseStarted":   [{urls: ["*://mochi.test/*"]}],
    "onCompleted":         [{urls: ["*://mochi.test/*"]}, ["responseHeaders"]],
    "onErrorOccurred":     [{urls: ["*://mochi.test/*"]}],
  };

  let extension = makeExtension(events);
  yield extension.startup();
  let authInfo = {
    username: "testuser",
    password: "testpass",
  };
  let expect = {
    "authenticate.sjs": {
      type: "xmlhttprequest",
      events: ["onBeforeRequest", "onBeforeSendHeaders", "onSendHeaders", "onHeadersReceived", "onResponseStarted", "onCompleted"],
    },
  };
  // expecting origin == undefined
  extension.sendMessage("set-expected", {expect, origin: location.href});
  yield extension.awaitMessage("continue");

  yield testXHR(`${baseUrl}?realm=webRequest_auth&user=${authInfo.username}&pass=${authInfo.password}`);

  yield extension.awaitMessage("done");
  yield extension.unload();
});

add_task(function* test_webRequest_cached_credentials2() {
  let authCredentials = {
    username: "testuser",
    password: "testpass",
  };
  let ex1 = getAuthHandler();
  yield ex1.startup();

  yield testXHR(`${baseUrl}?realm=webRequest_auth&user=${authCredentials.username}&pass=${authCredentials.password}`);

  yield ex1.awaitMessage("onCompleted");
  yield ex1.unload();
});

add_task(function* test_webRequest_window() {
  let authCredentials = {
    username: "testuser",
    password: "testpass",
  };
  let ex1 = getAuthHandler();
  yield ex1.startup();

  let win = window.open(`${baseUrl}?realm=test_webRequest_window&user=${authCredentials.username}&pass=${authCredentials.password}`);

  yield ex1.awaitMessage("onCompleted");
  yield ex1.unload();
  win.close();
});

add_task(function* test_webRequest_auth_cancelled() {
  let authCredentials = {
    username: "testuser_canceled",
    password: "testpass_canceled",
  };
  let ex1 = getAuthHandler({authCredentials});
  yield ex1.startup();
  let ex2 = getAuthHandler({cancel: true});
  yield ex2.startup();

  yield Assert.rejects(testXHR(`${baseUrl}?realm=test_webRequest_auth_cancelled&user=${authCredentials.username}&pass=${authCredentials.password}`), "caught rejected xhr");

  yield Promise.all([
    ex1.awaitMessage("onAuthRequired"),
    ex2.awaitMessage("onAuthRequired"),
    ex1.awaitMessage("onErrorOccurred"),
    ex2.awaitMessage("onErrorOccurred"),
  ]);
  yield ex1.unload();
  yield ex2.unload();
});

add_task(function* test_webRequest_auth_nonblocking() {
  // The first listener handles the auth request, the second listener
  // is a non-blocking listener and cannot respond but will get the call.
  let authCredentials = {
    username: "foobar",
    password: "testpass",
  };
  let handlingExt = getAuthHandler({authCredentials});
  yield handlingExt.startup();
  let extension = getAuthHandler({}, false);
  yield extension.startup();

  yield testXHR(`${baseUrl}?realm=webRequest_auth_nonblocking&user=${authCredentials.username}&pass=${authCredentials.password}`);

  yield Promise.all([
    extension.awaitMessage("onAuthRequired"),
    extension.awaitMessage("onCompleted"),
    handlingExt.awaitMessage("onAuthRequired"),
    handlingExt.awaitMessage("onCompleted"),
  ]);
  yield extension.unload();
  yield handlingExt.unload();
});


add_task(function* test_webRequest_auth_blocking_noreturn() {
  // The first listener is blocking but doesn't return anything.  The second
  // listener cancels the request.
  let ext = getAuthHandler();
  yield ext.startup();
  let canceler = getAuthHandler({cancel: true});
  yield canceler.startup();

  yield Assert.rejects(testXHR(`${baseUrl}?realm=auth_blocking_noreturn&user=auth_blocking_noreturn&pass=auth_blocking_noreturn`), "caught rejected xhr");

  yield Promise.all([
    ext.awaitMessage("onAuthRequired"),
    ext.awaitMessage("onErrorOccurred"),
    canceler.awaitMessage("onAuthRequired"),
    canceler.awaitMessage("onErrorOccurred"),
  ]);
  yield ext.unload();
  yield canceler.unload();
});

add_task(function* test_webRequest_auth_nonblocking_forwardAuthProvider() {
  // The chrome script sets up a default auth handler on the channel, the
  // extension does not return anything in the authRequred call.  We should
  // get the call in the extension first, then in the chrome code where we
  // cancel the request to avoid dealing with the prompt dialog here.  The test
  // is to ensure that WebRequest calls the previous notificationCallbacks
  // if the authorization is not handled by the onAuthRequired handler.

  let chromeScript = SpecialPowers.loadChromeScript(() => {
    const {classes: Cc, interfaces: Ci, utils: Cu, results: Cr} = Components;
    Cu.import("resource://gre/modules/Services.jsm");
    Cu.import("resource://gre/modules/XPCOMUtils.jsm");

    let observer = channel => {
      if (!(channel instanceof Ci.nsIHttpChannel && channel.URI.host === "mochi.test")) {
        return;
      }
      Services.obs.removeObserver(observer, "http-on-modify-request");
      channel.notificationCallbacks = {
        QueryInterface: XPCOMUtils.generateQI([Ci.nsIInterfaceRequestor,
                                               Ci.nsIAuthPromptProvider,
                                               Ci.nsIAuthPrompt2]),
        getInterface: XPCOMUtils.generateQI([Ci.nsIAuthPromptProvider,
                                             Ci.nsIAuthPrompt2]),
        promptAuth(channel, level, authInfo) {
          throw Cr.NS_ERROR_NO_INTERFACE;
        },
        getAuthPrompt(reason, iid) {
          return this;
        },
        asyncPromptAuth(channel, callback, context, level, authInfo) {
          // We just cancel here, we're only ensuring that non-webrequest
          // notificationcallbacks get called if webrequest doesn't handle it.
          Promise.resolve().then(() => {
            callback.onAuthCancelled(context, false);
            channel.cancel(Cr.NS_BINDING_ABORTED);
            sendAsyncMessage("callback-complete");
          });
        },
      };
    };
    Services.obs.addObserver(observer, "http-on-modify-request", false);
    sendAsyncMessage("chrome-ready");
  });
  yield chromeScript.promiseOneMessage("chrome-ready");
  let callbackComplete = chromeScript.promiseOneMessage("callback-complete");

  let handlingExt = getAuthHandler();
  yield handlingExt.startup();

  yield Assert.rejects(testXHR(`${baseUrl}?realm=auth_nonblocking_forwardAuth&user=auth_nonblocking_forwardAuth&pass=auth_nonblocking_forwardAuth`), "caught rejected xhr");

  yield callbackComplete;
  yield handlingExt.awaitMessage("onAuthRequired");
  // We expect onErrorOccurred because the "default" authprompt above cancelled
  // the auth request to avoid a dialog.
  yield handlingExt.awaitMessage("onErrorOccurred");
  yield handlingExt.unload();
  chromeScript.destroy();
});

add_task(function* test_webRequest_auth_nonblocking_forwardAuthPrompt2() {
  // The chrome script sets up a default auth handler on the channel, the
  // extension does not return anything in the authRequred call.  We should
  // get the call in the extension first, then in the chrome code where we
  // cancel the request to avoid dealing with the prompt dialog here.  The test
  // is to ensure that WebRequest calls the previous notificationCallbacks
  // if the authorization is not handled by the onAuthRequired handler.

  let chromeScript = SpecialPowers.loadChromeScript(() => {
    const {classes: Cc, interfaces: Ci, utils: Cu, results: Cr} = Components;
    Cu.import("resource://gre/modules/Services.jsm");
    Cu.import("resource://gre/modules/XPCOMUtils.jsm");

    let observer = channel => {
      if (!(channel instanceof Ci.nsIHttpChannel && channel.URI.host === "mochi.test")) {
        return;
      }
      Services.obs.removeObserver(observer, "http-on-modify-request");
      channel.notificationCallbacks = {
        QueryInterface: XPCOMUtils.generateQI([Ci.nsIInterfaceRequestor,
                                               Ci.nsIAuthPrompt2]),
        getInterface: XPCOMUtils.generateQI([Ci.nsIAuthPrompt2]),
        promptAuth(channel, level, authInfo) {
          throw Cr.NS_ERROR_NO_INTERFACE;
        },
        asyncPromptAuth(channel, callback, context, level, authInfo) {
          // We just cancel here, we're only ensuring that non-webrequest
          // notificationcallbacks get called if webrequest doesn't handle it.
          Promise.resolve().then(() => {
            channel.cancel(Cr.NS_BINDING_ABORTED);
            sendAsyncMessage("callback-complete");
          });
        },
      };
    };
    Services.obs.addObserver(observer, "http-on-modify-request", false);
    sendAsyncMessage("chrome-ready");
  });
  yield chromeScript.promiseOneMessage("chrome-ready");
  let callbackComplete = chromeScript.promiseOneMessage("callback-complete");

  let handlingExt = getAuthHandler();
  yield handlingExt.startup();

  yield Assert.rejects(testXHR(`${baseUrl}?realm=auth_nonblocking_forwardAuthPromptProvider&user=auth_nonblocking_forwardAuth&pass=auth_nonblocking_forwardAuth`), "caught rejected xhr");

  yield callbackComplete;
  yield handlingExt.awaitMessage("onAuthRequired");
  // We expect onErrorOccurred because the "default" authprompt above cancelled
  // the auth request to avoid a dialog.
  yield handlingExt.awaitMessage("onErrorOccurred");
  yield handlingExt.unload();
  chromeScript.destroy();
});

add_task(function* test_webRequest_duelingAuth() {
  let exNone = getAuthHandler();
  yield exNone.startup();
  let authCredentials = {
    username: "testuser_da1",
    password: "testpass_da1",
  };
  let ex1 = getAuthHandler({authCredentials});
  yield ex1.startup();
  let exEmpty = getAuthHandler({});
  yield exEmpty.startup();
  let ex2 = getAuthHandler({authCredentials: {
    username: "testuser_da2",
    password: "testpass_da2",
  }});
  yield ex2.startup();

  // XHR should succeed since the first credentials win, and they are correct.
  yield testXHR(`${baseUrl}?realm=test_webRequest_duelingAuth&user=${authCredentials.username}&pass=${authCredentials.password}`);

  yield Promise.all([
    exNone.awaitMessage("onAuthRequired"),
    exNone.awaitMessage("onCompleted"),
    exEmpty.awaitMessage("onAuthRequired"),
    exEmpty.awaitMessage("onCompleted"),
    ex1.awaitMessage("onAuthRequired"),
    ex1.awaitMessage("onCompleted"),
    ex2.awaitMessage("onAuthRequired"),
    ex2.awaitMessage("onCompleted"),
  ]);
  yield Promise.all([
    exNone.unload(),
    exEmpty.unload(),
    ex1.unload(),
    ex2.unload(),
  ]);
});

add_task(function* test_webRequest_auth_proxy() {
  function background() {
    let proxyOk = false;
    browser.webRequest.onAuthRequired.addListener((details) => {
      browser.test.succeed(`handlingExt onAuthRequired called with ${details.requestId} ${details.url}`);
      if (details.isProxy) {
        browser.test.succeed("providing proxy authorization");
        proxyOk = true;
        return {authCredentials: {username: "puser", password: "ppass"}};
      }
      browser.test.assertTrue(proxyOk, "providing www authorization after proxy auth");
      browser.test.sendMessage("done");
      return {authCredentials: {username: "auser", password: "apass"}};
    }, {urls: ["*://mochi.test/*"]}, ["blocking"]);
  }

  let handlingExt = ExtensionTestUtils.loadExtension({
    manifest: {
      permissions: [
        "webRequest",
        "webRequestBlocking",
        "*://mochi.test/*",
      ],
    },
    background,
  });

  yield handlingExt.startup();

  yield testXHR(`${baseUrl}?realm=auth_proxy&user=auser&pass=apass&proxy_user=puser&proxy_pass=ppass`);

  yield handlingExt.awaitMessage("done");
  yield handlingExt.unload();
});
</script>
</head>
<body>
<div id="test">Authorization Test</div>

</body>
</html>
